feat: Add image embedding functionality and metadata extraction
- Implemented C_embed_image.py for embedding images using MobileNetV2 and storing embeddings in MongoDB.
- Created 1_ETL_image.py and 2_embed_image.py for unzipping images and embedding them, respectively.
- Enhanced ETL_preprocess_helper.py with functions for extracting metadata and generating embeddings.
- Updated database_helper.py for improved MongoDB operations.
- Refactored file_helper.py to include zip extraction and file management utilities.
- Added similarity_helper.py for handling image and text embeddings for recommendation systems.

+++ src/utils/file_helper.py
@@ -7,23 +7,95 @@ import pickle
 import shutil
 import numpy as np
 import pandas as pd
+import zipfile
+from enum import Enum
+from dataclasses import dataclass
+from typing import List, Optional
 
 from . import setup_helper as sh
 from . import data_helper as dh
 
 ##########
-def move_file(file: Path, target_folder: Path):
-    src = Path(file)
+def move_file(file: Path, target_folder: Path, source_folder: Path=None):
+    file = Path(file)
 
+    if file.is_absolute():
+        src = file
+
+    elif source_folder:
+        src = Path(source_folder) / file
+
+    else:
+        raise FileNotFoundError(
+            f"Relative filename '{file}' received but no source_folder was provided. "
+            "Provide source_folder OR pass absolute paths."
+        )    
+
+    if not src.exists():
+        raise FileNotFoundError(f"Source file not found: {src}")
+    
     target_folder.mkdir(parents=True, exist_ok=True)
-
     dst = Path(target_folder) / src.name
+
     if dst.exists():
         dst.unlink()
 
-    shutil.move(str(file), str(dst))
+    shutil.move(str(src), str(dst))
     return dst
 
+
+class ExtractStatus(Enum):
+    DONE = "done"
+    SKIPPED = "skipped"
+
+
+@dataclass(frozen=True)
+class ExtractResult:
+    status: ExtractStatus
+    # für Python 3.10+
+    # files: list[Path] | None = None
+    # folders: list[Path] | None = None
+
+    # für Python <3.10
+    files: Optional[List[Path]] = None
+    folders: Optional[List[Path]] = None
+
+
+def unzip_images(f_names, extract_dir):
+    # zip_file = Path(src_folder) / "images.zip"
+    
+    img_files, img_folders = list_files_and_folders(extract_dir)
+    
+    if img_files or img_folders:
+        print(f"No files unzipped. Folder 'extract_dir' already contains {len(img_files)} files and {img_folders} folders .")        
+        return ExtractResult(status=ExtractStatus.SKIPPED) 
+    
+    for zip_file in f_names:
+        with zipfile.ZipFile(zip_file, 'r') as zip_ref:
+            zip_ref.extractall(extract_dir)
+
+        print(f"Unzipped file: {zip_file} to {extract_dir}")
+
+    files_unzipped, folders_unzipped = list_files_and_folders(extract_dir)
+
+    print(f"Unzipped {len(files_unzipped)} files and {folders_unzipped} folders to:", extract_dir)
+    
+    return ExtractResult(
+            status=ExtractStatus.DONE,
+            files=files_unzipped,
+            folders=folders_unzipped
+                    )   
+
+    # else:
+    #     print(f"No files unzipped. Folder 'extract_dir' already contains {len(img_files)} files and {img_folders} folders .")
+    #     return None, None 
+
+
+def list_files_and_folders(path):
+    files = [f for f in path.iterdir() if f.is_file()]
+    folders = [f for f in path.iterdir() if f.is_dir()]
+
+    return files, folders
 
 def save_pickle(file, path, folder=None):
     # now = datetime.now().strftime("%Y%m%d_%H%M%S")
@@ -42,6 +114,22 @@ def save_pickle(file, path, folder=None):
         print("[SUCCESS] File saved")
     except:
         print("[ERROR] Saving file as pkl.")
+
+
+def save_pickle2(obj, file_path):
+    # Ordner extrahieren
+    folder = os.path.dirname(file_path)
+
+    # Ordner erstellen falls nötig
+    if folder and not os.path.exists(folder):
+        os.makedirs(folder, exist_ok=True)
+
+    try:
+        with open(file_path, "wb") as f:
+            pickle.dump(obj, f)
+        print(f"[SUCCESS] File saved → {file_path}")
+    except Exception as e:
+        print("[ERROR] Saving file as pkl:", e)
 
 
 def merge_dfs(dict_df: dict):
@@ -96,23 +184,48 @@ def data_preview(f_names=None, folder=None):
     if not folder:
         print("No folder path provided")
 
-    df_dict = {}
-    for f in f_names:
-        f_path = os.path.join(folder, f"{f}.csv")
-        try:
-            df = pd.read_csv(f_path)
-            df_checked = column_check(df)
- 
-            # df = df.rename(columns={"Unnamed: 0": "id"}) 
-            # df["now"] = now_etl 
-            df_dict[f] = df_checked
-            sh.log_header("EDA RAW DATA")
+    df_dict = load_dfs(f_names, folder)
+
+    if len(df_dict) == 0:
+        return None 
+    
+    else:
+        for name, df in df_dict.items():
+            sh.log_header(f"EDA RAW DATA ({Path(name).name})")
             # print(f"\n{'='*30}\n--- EDA RAW DATA '{f}' ---") 
             print("SHAPE:\t", df.shape)
             print("INFO\n", sh.info_as_string(df))
             print(f"HEAD:\n{df.head(5)}\n")
 
-        except Exception as e:
-            print(f"⚠️ Loading {f} from {f_path}: Error occured:\n{e}")
+        return df_dict
 
-    return df_dict if len(df_dict) > 0 else None
+def load_dfs(f_names, folder=None):
+    df_dict = {}
+    
+    if isinstance(f_names, (str, Path)):
+        f_names = [f_names]
+        
+    for f in f_names:
+        name = Path(f).name
+
+        if folder:
+            f_path = Path(folder) / name
+        else:
+            f_path = name
+
+        print(f"[DEBUG] trying to load: {f_path}\n f={f}; folder={folder}")
+
+        try:  
+            df = pd.read_csv(f_path)            
+            df_dict[f_path] = df
+            
+        except Exception as e:
+            print(f"⚠️ Loading df '{name}' from {f_path}: Error occured:\n{e}")
+        
+    return df_dict
+
+
+
+
+
+
